requirements.txt
advent-of-code-data==2.0.4
annotated-types==0.7.0
aocd-example-parser==2023.12.20
beautifulsoup4==4.12.3
colorama==0.4.6
copier==9.4.1
dunamai==1.23.0
funcy==2.0
imageio==2.36.1
iniconfig==2.0.0
Jinja2==3.1.4
jinja2-ansible-filters==1.3.2
MarkupSafe==3.0.2
numpy==2.2.0
packaging==24.2
pathspec==0.12.1
Pebble==5.1.0
pillow==11.0.0
pluggy==1.5.0
plumbum==1.9.0
prompt-toolkit==3.0.36
pydantic==2.10.2
pydantic_core==2.27.1
Pygments==2.18.0
pytest==8.3.3
PyYAML==6.0.2
questionary==2.0.1
soupsieve==2.6
typing_extensions==4.12.2
urllib3==2.2.3
wcwidth==0.2.13
